refactor: set app mode and app zoom in html
this should help resolve many flickering/resizing issues
