+++ src/components/app/bottom-bar/bottom-bar.css
@@ -2,5 +2,5 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  font-size: 1rem;
+  font-size: 1em;
 }
