+++ src/components/media/media.css
@@ -19,7 +19,7 @@
 }
 
 .media-caption {
-  margin: 0;
+  margin: 0 0 1.5em;
   font-size: 75%;
   font-style: italic;
   color: var(--fg-muted);
